Create README.md
My third task as intern in PRODIGY INFOTECH.

Task about: Build a tool that assesses the strength of a password based on criteria such as length, presence of uppercase and lowercase, numbers, and special characters. Provide feedback to users on the password's strength. Develop a password complexity checker using python programming language.

Features

Password length: The programm checks if the define length criteria is meet by the password given by the user.

Presence of uppercase: The program checks for the presence of the uppercase in the password given by the user or not.

Presence of lowercase: The program checks for the presence of the lowercase in the password given by the user or not.

Presence of special character: The program checks for the presence of the special character defined in the program are used in tn password given by the user or not.

Presence of number: The program checks for the presence of the number between 0-9 used in the password given by the user or not.